Bugfix/sensor scan position integer (#1287)
* change sensorScanPosition from float to int

* change sensorScanPosition from float to int

* fix _FillValue to match and pass CI tests

* fix _FillValue to pass CI tests

* add new block to radiance vars MetaData to handle integers in addition to floats

* change sensorScanPosition to integer and eliminate duplicate of fieldOfViewNumber

* fix a few _FillValue and eliminate duplicative fieldOfViewNumber

* fix, again, the FillValue to be positive for these files

* add the sensor to field of view and range for Tbs

* Revert "add the sensor to field of view and range for Tbs"

This reverts commit d1b65a6656b4febc5a1da8b01c47e8b58d7e6c9a.

* update testoutput for amsua both in sensorScanPosition fix and valid_range

---------

+++ src/hdf5/amsr2_2ioda.py
@@ -149,7 +149,7 @@ def get_data(f, obs_data):
     # start at channel 5 as lowest frequencies are not included
     obs_data[('sensorChannelNumber', metaDataName)] = np.array(np.arange(nchans), dtype='int32')
     k = 'sensorScanPosition'
-    obs_data[(k, metaDataName)] = np.tile(np.arange(nbeam_pos, dtype='float32')+1, (nscans, 1)).flatten()
+    obs_data[(k, metaDataName)] = np.tile(np.arange(nbeam_pos, dtype='int32')+1, (nscans, 1)).flatten()
     k = 'sensorZenithAngle'   # ~55.2 incidence angle
     obs_data[(k, metaDataName)] = np.array(f['Earth Incidence'], dtype='float32').flatten()
     instr_scan_ang = obs_data[(k, metaDataName)]
